# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP V2 API",
        "description": "Comprehensive ERP System API Documentation with Multi-tenancy and Role-based Access Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/subscriptions": {
            "get": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Get all subscriptions",
                "operationId": "35cad7b3411a18e5aa61df44434bb2f7",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Create a new subscription",
                "operationId": "f0596984bf1b4c4427e9146fe3649154",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Get subscription by ID",
                "operationId": "57e14c8f4d4c20e65c910dffb03566f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Update subscription",
                "operationId": "24c874a70f93de774a7294ddea27501f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Delete subscription",
                "operationId": "f19ba63504cde0835137300fe6a85015",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subscription deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/subscriptions/subscriptions/{id}/change-status": {
            "patch": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Change subscription status",
                "operationId": "81690d1efd7164bb8508f4b39f380128",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status changed"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/subscriptions/subscriptions/subscribe-user": {
            "post": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Subscribe a user to a plan",
                "operationId": "4c4aca89a46851632643dab40d5c7c0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubscribeUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User subscribed"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/subscriptions/subscriptions/{id}/renew": {
            "post": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Renew subscription",
                "operationId": "b43c5dce393fd1968e563ebb7a0026e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription renewed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/subscriptions/subscriptions/{id}/cancel": {
            "post": {
                "tags": [
                    "Admin/Subscriptions"
                ],
                "summary": "Cancel subscription",
                "operationId": "2f36368eea41064a2515683fe32e94fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnsubscribeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription cancelled"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreSubscriptionRequest": {
                "required": [
                    "user_id",
                    "plan_id",
                    "start_date",
                    "end_date",
                    "status",
                    "price_paid",
                    "billing_cycle"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-01"
                    },
                    "trial_ends_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "expired",
                            "cancelled",
                            "trial"
                        ]
                    },
                    "auto_renew": {
                        "type": "boolean",
                        "example": true
                    },
                    "price_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "billing_cycle": {
                        "type": "string",
                        "example": "monthly",
                        "enum": [
                            "monthly",
                            "yearly",
                            "lifetime"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateSubscriptionRequest": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-01"
                    },
                    "trial_ends_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "expired",
                            "cancelled",
                            "trial"
                        ]
                    },
                    "auto_renew": {
                        "type": "boolean",
                        "example": true
                    },
                    "price_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "billing_cycle": {
                        "type": "string",
                        "example": "monthly",
                        "enum": [
                            "monthly",
                            "yearly",
                            "lifetime"
                        ]
                    },
                    "cancellation_reason": {
                        "type": "string",
                        "example": "Changed my mind",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UnsubscribeRequest": {
                "properties": {
                    "reason": {
                        "type": "string",
                        "example": "Too expensive",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ChangeStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "expired",
                            "cancelled",
                            "trial"
                        ]
                    }
                },
                "type": "object"
            },
            "SubscribeUserRequest": {
                "required": [
                    "user_id",
                    "plan_id"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_trial": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}